requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
cryptography==2.6.1
future==0.17.1
numpy==1.16.2
pdf2image==1.4.2
Pillow==5.4.1
pycparser==2.19
pytesseract==0.2.6
python-telegram-bot==12.0.0b1
schedule==0.6.0
six==1.12.0
tornado==6.0.1
